+++ src/utils/ui_overlay.hpp
@@ -9,6 +9,6 @@ namespace ui_overlay {
 	void draw_time(uint32_t time);
 	void draw_game_info(std::string info_text, uint32_t time);
 	void draw_points(uint32_t points);
-	void draw_fps(float &ms_start, float &ms_end);
+	void draw_fps(uint32_t &ms_start, uint32_t &ms_end);
 	void draw_game_over(bool win_game);
 }
